more ego filter ui implementation
